This is the final commit
AS this is the final commit, I have changed a few things and modified the results section a bit. I wanted to make my results look unique and cool. So I added a bit of color scheme here, wel it is not too tough. But looks really good. And there were a few files that were created due to the VS Community 2022. And I also updated this software yesterday, so there are a few more extra files